feat: moved function from utils to correct place, commented functionalities in different steps
